Use form mixin for add field change handler
