+++ js/components/forms/financial.js
@@ -1,8 +1,11 @@
+import FormMixin from '../../mixins/form'
 import optionsinput from '../options_input'
 import textinput from '../text_input'
 
 export default {
   name: 'financial',
+
+  mixins: [FormMixin],
 
   components: {
     optionsinput,
@@ -24,18 +27,5 @@ export default {
     return {
       funding_type
     }
-  },
-
-  mounted: function () {
-    this.$root.$on('field-change', this.handleFieldChange)
-  },
-
-  methods: {
-    handleFieldChange: function (event) {
-      const { value, name } = event
-      if (typeof this[name] !== undefined) {
-        this[name] = value
-      }
-    },
   }
 }
